TIP-221: Move GroupFactory
