+++ src/Pim/Bundle/CatalogBundle/Resources/config/factories.yml
@@ -2,7 +2,7 @@ parameters:
     pim_catalog.factory.family.class:                Pim\Component\Catalog\Factory\FamilyFactory
     pim_catalog.factory.attribute_requirement.class: Pim\Bundle\CatalogBundle\Factory\AttributeRequirementFactory
     pim_catalog.factory.metric.class:                Pim\Component\Catalog\Factory\MetricFactory
-    pim_catalog.factory.group.class:                 Pim\Bundle\CatalogBundle\Factory\GroupFactory
+    pim_catalog.factory.group.class:                 Pim\Component\Catalog\Factory\GroupFactory
     pim_catalog.factory.attribute.class:             Pim\Bundle\CatalogBundle\Factory\AttributeFactory
     pim_catalog.factory.group_type.class:            Pim\Component\Catalog\Factory\GroupTypeFactory
     pim_catalog.factory.currency.class:              Pim\Component\Catalog\Factory\CurrencyFactory
